--- a/thicket/ensemble.py
+++ b/thicket/ensemble.py
@@ -216,9 +216,9 @@ class Ensemble:
             combined_th.profile = [new_mappings[prf] for prf in combined_th.profile]
             profile_mapping_cp = combined_th.profile_mapping.copy()
             for k, v in profile_mapping_cp.items():
-                combined_th.profile_mapping[new_mappings[k]] = (
-                    combined_th.profile_mapping.pop(k)
-                )
+                combined_th.profile_mapping[
+                    new_mappings[k]
+                ] = combined_th.profile_mapping.pop(k)
             combined_th.performance_cols = helpers._get_perf_columns(
                 combined_th.dataframe
             )