--- a/thicket/ensemble.py
+++ b/thicket/ensemble.py
@@ -35,7 +35,18 @@ class Ensemble:
                 (list): list of Thicket objects
         """
 
-        def _update_graph_and_df(thicket, old_to_new_dict, union_graph, _debug=False):
+        def _replace_graph_df_nodes(thicket, old_to_new, union_graph, _debug=False):
+            """Replace the node objects in the graph and DataFrame of a Thicket object from the result of graph.union(). 
+            
+            Arguments:
+                thicket (Thicket): Thicket object to modify
+                old_to_new (dict): dictionary mapping old node objects to new node objects
+                union_graph (hatchet.Graph): unified graph
+                _debug (bool): whether to print debug statements
+            
+            Returns:
+                (Thicket): modified Thicket object
+            """
             if _debug:
                 print(f"Graph: {thicket.profile_mapping}")
             thicket.graph = union_graph
@@ -47,13 +58,13 @@ class Ensemble:
             replace_dict = {}
             for node in thicket.dataframe["node"]:
                 node_id = id(node)
-                if node_id in old_to_new_dict:
+                if node_id in old_to_new:
                     if _debug:
                         print(
-                            f"Updating node: {node_id} {node} {hash(node)}\n\t-> {id(old_to_new_dict[node_id])} {old_to_new_dict[node_id]} {hash(old_to_new_dict[node_id])}"
+                            f"Updating node: {node_id} {node} {hash(node)}\n\t-> {id(old_to_new[node_id])} {old_to_new[node_id]} {hash(old_to_new[node_id])}"
                         )
-                    check_same_frame(node, old_to_new_dict[node_id])
-                    replace_dict[node] = old_to_new_dict[node_id]
+                    check_same_frame(node, old_to_new[node_id])
+                    replace_dict[node] = old_to_new[node_id]
             # Replace in one-op for optimization
             thicket.dataframe["node"] = thicket.dataframe["node"].replace(replace_dict)
             thicket.dataframe = thicket.dataframe.set_index(idx_names)
@@ -69,7 +80,7 @@ class Ensemble:
             temp_dict = {}
             union_graph = union_graph.union(_thickets[i + 1].graph, temp_dict)
             for j in range(i + 2):
-                _thickets[j] = _update_graph_and_df(
+                _thickets[j] = _replace_graph_df_nodes(
                     _thickets[j], temp_dict, union_graph, _debug=False
                 )
         for i in range(len(_thickets)):