--- a/thicket/helpers.py
+++ b/thicket/helpers.py
@@ -3,8 +3,6 @@
 #
 # SPDX-License-Identifier: MIT
 
-import copy
-
 import pandas as pd