--- a/thicket/tests/test_thicket.py
+++ b/thicket/tests/test_thicket.py
@@ -3,7 +3,6 @@
 #
 # SPDX-License-Identifier: MIT
 
-import copy
 import re
 
 import pytest