… thicket/tests/conftest.py
… thicket/thicket.py