@michaelmckinsey1 2024-01-10
Resolves #16